new i18n and translator file. [es6]
